+++ components-ui/app/src/tol-ui/src/general/Icon.tsx
@@ -5,6 +5,7 @@ SPDX-License-Identifier: MIT
 */
 
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import type { IconProp, SizeProp } from "@fortawesome/fontawesome-svg-core";
 import { HoverOverlay } from "./HoverOverlay";
 
 
@@ -52,8 +53,14 @@ export interface PIcon {
 export function Icon(props: PIcon) {
   const { icon, size, colour, config = "solid", className, onClick, tooltip, testid } = props;
 
-  // @ts-ignore
-  const IconContents = <FontAwesomeIcon icon={`fa-${config} fa-${icon}`} size={size} color={colour} />;
+  const prefix = config === "brands" ? "fab" : config === "regular" ? "far" : "fas";
+  const IconContents = (
+    <FontAwesomeIcon
+      icon={[prefix, icon] as IconProp}
+      size={size as SizeProp}
+      color={colour}
+    />
+  );
 
   return (
     <span